Moving attribute helper functions to a separate file and adding mergeTabsterProps().
